Adds PAT scope information to resolve git clone error.
Having PAT setup without full `repo` scope selected caused 403 error on cloning. Even with specific sub-settings selected under the `repo` scope the 403 still occurred.  The scope requirement is mentioned in the "Creating a personal access token" documentation, but this can help serve as a reminder for users with a PAT created who may not have selected the full `repo` scope.